requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.2
Django==4.1.7
django-cors-headers==3.14.0
django-extensions==3.2.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
mongoengine==0.27.0
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
pymongo==3.12.3
python-decouple==3.8
pytz==2022.7.1
sqlparse==0.2.4
tzdata==2022.7
